dropbear: Work around bug in buildroot-2015.08-rc1
As of buildroot-2015.08-rc1 /etc/dropbear is a broken symlink
to /var/run/dropbear (which in turn is a symlink to /run/dropbear).
We replace the symlink (or the existing directory in case of
incremental builds) with a real directory so that dropbear can
store the generated host keys when the first client connects.